backend/requirements.txt
# Django and Django REST Framework
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
django-filter==23.3

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Authentication
djangorestframework-simplejwt==5.3.0

# API Documentation
drf-spectacular==0.26.5

# Static files
whitenoise==6.6.0

# Error tracking
sentry-sdk==1.38.0

# Production server
gunicorn==21.2.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Code quality
flake8==6.1.0
black==23.11.0
isort==5.12.0

# Environment variables
python-dotenv==1.0.0

# Utilities
python-dateutil==2.8.2
Pillow==10.1.0  # For image handling in lab results, etc.
